+++ recalc-vscode/src/frontend/index.ts
@@ -1,1 +1,10 @@
-console.log("Hello, TypeScript!");
+import { WebviewApi } from 'vscode-webview';
+
+const vscode: WebviewApi<never> = acquireVsCodeApi();
+
+window.addEventListener('message',
+  (x: string) =>
+    console.log(`GOT MESSAGE: ${x}`)
+);
+
+vscode.postMessage({method: "hi", params: {origin: "frontend", destination: "haskell"}});