+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for recalc
 
-## 0.3.1.3 -- 2025-02-05
+## 0.4.0.0 -- 2025-02-07
 
 * Initial version: project skeleton for recalculation engine + test suite
 * First implementation: minimal recalculation engine for generic language
@@ -11,3 +11,4 @@
 * Add Syntax for the term language (simple dependently typed lambda calculus).
 * Add support for operators with precedences.
 * Prepare TypeScript code generation.
+* Add server implementation that is used as backend by the TypeScript frontend.