+++ recalc-vscode/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Revision history for recalc-vscode
 
-## 0.0.1 -- 2025-02-06
+## 0.0.2 -- 2025-02-06
 
 * Initial version: create project structure for a custom editor.
+* Add generic rpc implementation, client implementation and logging facilities.